Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main(){
-    int a, b;
-    int res;
+    float a, b;
+    float res;
     char c;
-    scanf("%d %d %c", &a, &b, &c);
+    scanf("%f %f %c", &a, &b, &c);
     switch(c){
         case '+': 
             res = a+b;
@@ -18,6 +18,6 @@ int main(){
             res = a/b;
             break;
     }
-    printf("%d", res);
+    printf("%.2f", res);
     return 0;
 }
